+++ next.config.js
@@ -1,4 +1,6 @@
-module.exports = {
+const withCSS = require('@zeit/next-css')
+
+module.exports = withCSS({
   useFileSystemPublicRoutes: false,
   assetPrefix: process.env.BASE_URL || '',
-}
+})